requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
googlemaps==4.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
mysql-connector-python==9.0.0
packaging==24.1
pillow==10.4.0
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
